api/requirements-vercel.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
python-multipart==0.0.6
hana-ml>=2.17.24011800
torch>=2.0.0
numpy>=1.24.0
transformers>=4.35.0
python-jose==3.3.0
PyJWT>=2.7.0
passlib==1.7.4
httpx==0.26.0
prometheus-client==0.19.0
python-dotenv==1.0.0
starlette==0.31.1
jinja2==3.1.2
langchain>=0.0.267
langchain-community>=0.0.6
langchain-core>=0.0.12
langchain-experimental>=0.0.5
langchain-text-splitters>=0.0.1
cryptography>=41.0.4
ipython<9.0
# GPU dependencies intentionally excluded for Vercel deployment